ids changed to integer, upload file fixed

+++ src/main/java/com/epam/resourceservice/dto/DeletedFilesDTO.java
@@ -13,5 +13,5 @@ import java.util.List;
 @Builder
 public class DeletedFilesDTO {
 
-    List<Long> ids;
+    List<Integer> ids;
 }
